Move email validation to account setup
