+++ feature/account/server/settings/src/main/kotlin/app/k9mail/feature/account/server/settings/ui/common/mapper/ValidationErrorStringMapper.kt
@@ -3,7 +3,6 @@ package app.k9mail.feature.account.server.settings.ui.common.mapper
 import android.content.res.Resources
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.feature.account.server.settings.R
-import app.k9mail.feature.account.server.settings.domain.usecase.ValidateEmailAddress.ValidateEmailAddressError
 import app.k9mail.feature.account.server.settings.domain.usecase.ValidateImapPrefix.ValidateImapPrefixError
 import app.k9mail.feature.account.server.settings.domain.usecase.ValidatePassword.ValidatePasswordError
 import app.k9mail.feature.account.server.settings.domain.usecase.ValidatePort.ValidatePortError
@@ -12,25 +11,12 @@ import app.k9mail.feature.account.server.settings.domain.usecase.ValidateUsernam
 
 fun ValidationError.toResourceString(resources: Resources): String {
     return when (this) {
-        is ValidateEmailAddressError -> toEmailAddressErrorString(resources)
         is ValidateServerError -> toServerErrorString(resources)
         is ValidatePortError -> toPortErrorString(resources)
         is ValidateUsernameError -> toUsernameErrorString(resources)
         is ValidatePasswordError -> toPasswordErrorString(resources)
         is ValidateImapPrefixError -> toImapPrefixErrorString(resources)
         else -> throw IllegalArgumentException("Unknown error: $this")
-    }
-}
-
-private fun ValidateEmailAddressError.toEmailAddressErrorString(resources: Resources): String {
-    return when (this) {
-        is ValidateEmailAddressError.EmptyEmailAddress -> resources.getString(
-            R.string.account_server_settings_validation_error_email_address_required,
-        )
-
-        is ValidateEmailAddressError.InvalidEmailAddress -> resources.getString(
-            R.string.account_server_settings_validation_error_email_address_invalid,
-        )
     }
 }
