+++ feature/account/server/settings/src/main/res/values/strings.xml
@@ -29,8 +29,6 @@
 
     <string name="account_server_settings_outgoing_top_bar_title">Outgoing server settings</string>
 
-    <string name="account_server_settings_validation_error_email_address_required">Email address is required.</string>
-    <string name="account_server_settings_validation_error_email_address_invalid">Email address is invalid.</string>
     <string name="account_server_settings_validation_error_server_required">Server name is required.</string>
     <string name="account_server_settings_validation_error_port_required">Port is required.</string>
     <string name="account_server_settings_validation_error_port_invalid">Port is invalid (must be 1–65535).</string>
