+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/domain/DomainContract.kt
@@ -22,6 +22,10 @@ interface DomainContract {
             ): String
         }
 
+        fun interface ValidateEmailAddress {
+            fun execute(emailAddress: String): ValidationResult
+        }
+
         fun interface ValidateConfigurationApproval {
             fun execute(isApproved: Boolean?, isAutoDiscoveryTrusted: Boolean?): ValidationResult
         }
