+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/domain/usecase/ValidateEmailAddress.kt
@@ -1,8 +1,8 @@
-package app.k9mail.feature.account.server.settings.domain.usecase
+package app.k9mail.feature.account.setup.domain.usecase
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
-import app.k9mail.feature.account.server.settings.domain.ServerSettingsDomainContract.UseCase
+import app.k9mail.feature.account.setup.domain.DomainContract.UseCase
 
 class ValidateEmailAddress : UseCase.ValidateEmailAddress {
