+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/AccountAutoDiscoveryValidator.kt
@@ -1,14 +1,14 @@
 package app.k9mail.feature.account.setup.ui.autodiscovery
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
-import app.k9mail.feature.account.server.settings.domain.usecase.ValidateEmailAddress
 import app.k9mail.feature.account.server.settings.domain.usecase.ValidatePassword
 import app.k9mail.feature.account.setup.domain.DomainContract.UseCase
 import app.k9mail.feature.account.setup.domain.usecase.ValidateConfigurationApproval
+import app.k9mail.feature.account.setup.domain.usecase.ValidateEmailAddress
 import app.k9mail.feature.account.server.settings.domain.ServerSettingsDomainContract.UseCase as ServerSettingsUseCase
 
 internal class AccountAutoDiscoveryValidator(
-    private val emailAddressValidator: ServerSettingsUseCase.ValidateEmailAddress = ValidateEmailAddress(),
+    private val emailAddressValidator: UseCase.ValidateEmailAddress = ValidateEmailAddress(),
     private val passwordValidator: ServerSettingsUseCase.ValidatePassword = ValidatePassword(),
     private val configurationApprovalValidator: UseCase.ValidateConfigurationApproval = ValidateConfigurationApproval(),
 ) : AccountAutoDiscoveryContract.Validator {
