+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/AutoDiscoveryStringMapper.kt
@@ -5,6 +5,7 @@ import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.feature.account.setup.R
 import app.k9mail.feature.account.setup.domain.entity.AutoDiscoveryConnectionSecurity
 import app.k9mail.feature.account.setup.domain.usecase.ValidateConfigurationApproval
+import app.k9mail.feature.account.setup.domain.usecase.ValidateEmailAddress
 
 internal fun AutoDiscoveryConnectionSecurity.toResourceString(resources: Resources): String {
     return when (this) {
@@ -27,11 +28,24 @@ internal fun AccountAutoDiscoveryContract.Error.toResourceString(resources: Reso
 
 internal fun ValidationError.toResourceString(resources: Resources): String {
     return when (this) {
+        is ValidateEmailAddress.ValidateEmailAddressError -> toEmailAddressErrorString(resources)
         is ValidateConfigurationApproval.ValidateConfigurationApprovalError -> toConfigurationApprovalErrorString(
             resources,
         )
 
         else -> throw IllegalArgumentException("Unknown error: $this")
+    }
+}
+
+private fun ValidateEmailAddress.ValidateEmailAddressError.toEmailAddressErrorString(resources: Resources): String {
+    return when (this) {
+        is ValidateEmailAddress.ValidateEmailAddressError.EmptyEmailAddress -> resources.getString(
+            R.string.account_setup_auto_discovery_validation_error_email_address_required,
+        )
+
+        is ValidateEmailAddress.ValidateEmailAddressError.InvalidEmailAddress -> resources.getString(
+            R.string.account_setup_auto_discovery_validation_error_email_address_invalid,
+        )
     }
 }
