+++ feature/account/setup/src/main/res/values/strings.xml
@@ -6,6 +6,9 @@
     <string name="account_setup_error_network">Network </string>
     <string name="account_setup_error_unknown">Unknown error</string>
 
+    <string name="account_setup_auto_discovery_validation_error_email_address_required">Email address is required.</string>
+    <string name="account_setup_auto_discovery_validation_error_email_address_invalid">Email address is invalid.</string>
+
     <string name="account_setup_auto_discovery_connection_security_start_tls">StartTLS</string>
     <string name="account_setup_auto_discovery_connection_security_ssl">SSL/TLS</string>
     <string name="account_setup_auto_discovery_loading_message">Finding email details</string>
